Updated Mednafen PSX core to 0.9.26

+++ PSX/Core/mednafen/tests.cpp
@@ -1,3 +1,5 @@
+// DO NOT REMOVE/DISABLE THESE MATH AND COMPILER SANITY TESTS.  THEY EXIST FOR A REASON.
+
 /* Mednafen - Multi-system Emulator
  *
  * This program is free software; you can redistribute it and/or modify
@@ -39,7 +41,7 @@ typedef struct
  int32 mostnegresult;
 } MathTestEntry;
 
-#define ADD_MTE(_bits) { _bits, ((uint32)1 << _bits) - 1, (uint32)1 << (_bits - 1), 0 - ((uint32)1 << (_bits - 1)) }
+#define ADD_MTE(_bits) { _bits, ((uint32)1 << _bits) - 1, (uint32)1 << (_bits - 1), (int32)(0 - ((uint32)1 << (_bits - 1))) }
 
 MathTestEntry math_test_vals[] =
 {
